add slate-react
